Support 29bit identifier for ISO-TP.

+++ src/ecu_lua_script.h
@@ -22,7 +22,7 @@ constexpr char READ_SEED[] = "Seed";
 constexpr char RAW_TABLE[] = "Raw";
 constexpr char J1939_SOURCE_ADDRESS_FIELD[] = "J1939SourceAddress";
 constexpr char J1939_PGN_TABLE[] = "PGNs";
-constexpr uint16_t DEFAULT_BROADCAST_ADDR = 0x7DF;
+constexpr uint32_t DEFAULT_BROADCAST_ADDR = 0x7DF;
 
 class EcuLuaScript
 {
@@ -35,9 +35,9 @@ public:
     EcuLuaScript& operator =(EcuLuaScript&& orig) noexcept;
     virtual ~EcuLuaScript() = default;
 
-    std::uint16_t getRequestId() const;
-    std::uint16_t getResponseId() const;
-    std::uint16_t getBroadcastId() const;
+    std::uint32_t getRequestId() const;
+    std::uint32_t getResponseId() const;
+    std::uint32_t getBroadcastId() const;
     std::uint8_t getJ1939SourceAddress() const;
     std::string getSeed(std::uint8_t identifier);
     std::string getDataByIdentifier(const std::string& identifier);
@@ -68,9 +68,9 @@ private:
     std::string ecu_ident_;
     SessionController* pSessionCtrl_ = nullptr;
     IsoTpSender* pIsoTpSender_ = nullptr;
-    std::uint16_t requestId_;
-    std::uint16_t responseId_;
-    std::uint16_t broadcastId_ = DEFAULT_BROADCAST_ADDR;
+    std::uint32_t requestId_;
+    std::uint32_t responseId_;
+    std::uint32_t broadcastId_ = DEFAULT_BROADCAST_ADDR;
     std::uint8_t j1939SourceAddress_;
     std::mutex luaLock_;
 };
